Renamed bbq/test.rb to bbq/test_unit.rb
